[!!!][TASK] Rename `renderSlot` to `slot.render`
This anticipates the future new view helper `slot.has`, and adds a new folder
reserved for the slots usage.
