+++ Resources/Private/Layouts/Bootstrap/Bootstrap3/RadioButton.html
@@ -16,7 +16,7 @@
 						</f:else>
 					</f:if>
 					<span class="input-group-addon">
-						<formz:renderSlot slot="Field-{key}" />
+						<formz:slot.render slot="Field-{key}" />
 		            </span>
 					<f:if condition="{item.reverse}">
 						<label class="btn btn-default form-control" for="{fieldId}-{key}">{item.label}</label>
