+++ Resources/Private/Layouts/Foundation/Foundation5/RadioButton.html
@@ -14,7 +14,7 @@
                         <f:if condition="{item.reverse}">
                             <f:then>
                                 <span class="prefix">
-                                    <formz:renderSlot slot="Field-{key}" />
+                                    <formz:slot.render slot="Field-{key}" />
                                 </span>
                             </f:then>
                             <f:else>
@@ -30,7 +30,7 @@
                             </f:then>
                             <f:else>
                                 <span class="postfix">
-                                    <formz:renderSlot slot="Field-{key}" />
+                                    <formz:slot.render slot="Field-{key}" />
                                 </span>
                             </f:else>
                         </f:if>
