+++ Resources/Private/Templates/Bootstrap/Default.html
@@ -5,22 +5,22 @@
 <f:layout name="{layout}" />
 
 <f:section name="Label">
-    <formz:renderSlot slot="Label.Before" />
+    <formz:slot.render slot="Label.Before" />
 
     <label class="control-label" for="{fieldId}">{label}</label>
 
-    <formz:renderSlot slot="Label.After" />
+    <formz:slot.render slot="Label.After" />
 </f:section>
 
 <f:section name="Field">
-    <formz:renderSlot slot="Field" />
+    <formz:slot.render slot="Field" />
 </f:section>
 
 <f:section name="Feedback">
-    <formz:renderSlot slot="Feedback.Out.Before" />
+    <formz:slot.render slot="Feedback.Out.Before" />
 
     <span class="help-block" formz-field-feedback-container="{fieldName}">
-		<formz:renderSlot slot="Feedback.In.Before" />
+		<formz:slot.render slot="Feedback.In.Before" />
 
 		<span formz-field-feedback-list="{fieldName}">
 			<f:form.validationResults for="{formName}.{fieldName}">
@@ -48,10 +48,10 @@
             </f:form.validationResults>
 		</span>
 
-		<formz:renderSlot slot="Feedback.In.After" />
+		<formz:slot.render slot="Feedback.In.After" />
 	</span>
 
-    <formz:renderSlot slot="Feedback.Out.After" />
+    <formz:slot.render slot="Feedback.Out.After" />
 </f:section>
 
 </html>
