+++ Resources/Private/Templates/Default/Default.html
@@ -9,14 +9,14 @@
 </f:section>
 
 <f:section name="Field">
-    <formz:renderSlot slot="Field" />
+    <formz:slot.render slot="Field" />
 </f:section>
 
 <f:section name="Feedback">
-	<formz:renderSlot slot="Feedback.Out.Before" />
+	<formz:slot.render slot="Feedback.Out.Before" />
 
 	<div formz-field-feedback-container="{fieldName}">
-		<formz:renderSlot slot="Feedback.In.Before" />
+		<formz:slot.render slot="Feedback.In.Before" />
 
 		<div formz-field-feedback-list="{fieldName}">
             <f:form.validationResults for="{formName}.{fieldName}">
@@ -44,10 +44,10 @@
             </f:form.validationResults>
 		</div>
 
-		<formz:renderSlot slot="Feedback.In.After" />
+		<formz:slot.render slot="Feedback.In.After" />
 	</div>
 
-	<formz:renderSlot slot="Feedback.Out.After" />
+	<formz:slot.render slot="Feedback.Out.After" />
 </f:section>
 
 </html>
