+++ Resources/Private/Templates/Foundation/Foundation5/Default.html
@@ -19,7 +19,7 @@
 </f:section>
 
 <f:section name="Field">
-	<formz:renderSlot slot="Field" />
+	<formz:slot.render slot="Field" />
 </f:section>
 
 <f:section name="Feedback">
@@ -50,10 +50,10 @@
 			</f:form.validationResults>
 		</div>
 
-		<formz:renderSlot slot="Feedback.In.After" />
+		<formz:slot.render slot="Feedback.In.After" />
 	</div>
 
-	<formz:renderSlot slot="Feedback.Out.After" />
+	<formz:slot.render slot="Feedback.Out.After" />
 </f:section>
 
 </html>
